+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 2019-05-03 v1.3.0-SNAPSHOT
  * Borrowed mapping of MARC to Instance logic from mod-data-loader. After parsing Records are mapped to Instances and saved in mod-inventory.
  * Added support for records coming for processing in json format.
+ * Applied logic for expanding parsed MARC records with additional custom fields (999 field)
 
 ## 2019-03-25 v1.2.1
  * Set required recordType field for the Record entity